# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  api_internal_handler.AuditLogEntry:
    properties:
      action:
        type: string
      actor_email:
        type: string
      created_at:
        type: string
      id:
        type: string
      ip:
        type: string
      metadata:
        description: JSON string
        type: string
      project_id:
        type: string
      result:
        type: string
      secret_hash:
        description: base64
        type: string
      token_id:
        type: string
      user_id:
        type: string
    type: object
  api_internal_handler.AuditLogListResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/api_internal_handler.AuditLogEntry'
        type: array
      meta:
        $ref: '#/definitions/api_internal_handler.Meta'
    type: object
...
  api_internal_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api_internal_handler.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: api.zenv.sh
info:
  contact: {}
  description: Zero-knowledge secret manager API. Stores and retrieves ciphertext.
    Never decrypts.
  title: zEnv API
  version: 0.1.0
paths:
  /audit-logs:
    get:
      description: Query audit log entries with filters. Requires project_id.
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Filter by action (e.g. secret.read)
        in: query
        name: action
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by result (success, denied, error)
        in: query
        name: result
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 50, max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_internal_handler.AuditLogListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
      security:
      - SessionAuth: []
      summary: List audit logs
      tags:
      - audit
  /audit-logs/drain:
    post:
      description: Force-flush all queued audit events from Redis to Postgres.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Drain audit queue
      tags:
      - audit
  /audit-logs/export:
    get:
      description: Export filtered audit log entries as a CSV file.
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by result
        in: query
        name: result
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV stream (Content-Disposition attachment)
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_internal_handler.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Export audit logs as CSV
      tags:
      - audit
...
securityDefinitions:
  BearerAuth:
    description: 'Service token: "Bearer ze_..."'
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"